+++ typescript/sdk/src/index.ts
@@ -1,4 +1,8 @@
-export { isAddressActive } from './contracts/contracts.js';
+export {
+  isAddressActive,
+  isContractAddress,
+  assertIsContractAddress,
+} from './contracts/contracts.js';
 export { MUTABLE_HOOK_TYPE } from './hook/types.js';
 export { MUTABLE_ISM_TYPE } from './ism/types.js';
 
@@ -122,7 +126,7 @@ export {
 } from './deploy/verify/types.js';
 export * as verificationUtils from './deploy/verify/utils.js';
 export { ZKSyncContractVerifier } from './deploy/verify/ZKSyncContractVerifier.js';
-export { executeWarpDeploy } from './deploy/warp.js';
+export { executeWarpDeploy, enrollCrossChainRouters } from './deploy/warp.js';
 export {
   SealevelIgpAdapter,
   SealevelOverheadIgpAdapter,
@@ -186,6 +190,8 @@ export {
 } from './hook/types.js';
 export { isHookCompatible } from './hook/utils.js';
 export { CosmosNativeIsmReader } from './ism/CosmosNativeIsmReader.js';
+export { CosmosNativeWarpModule } from './token/CosmosNativeWarpModule.js';
+export { CosmosNativeWarpRouteReader } from './token/CosmosNativeWarpRouteReader.js';
 export { EvmIsmReader } from './ism/EvmIsmReader.js';
 export { HyperlaneIsmFactory } from './ism/HyperlaneIsmFactory.js';
 export { BaseMetadataBuilder } from './ism/metadata/builder.js';
@@ -379,6 +385,7 @@ export {
   ViemProvider,
   ViemTransaction,
   ViemTransactionReceipt,
+  GroupedTransactions,
   ProtocolTypedTransaction,
 } from './providers/ProviderType.js';
 export {
@@ -403,6 +410,7 @@ export {
 export { CallData, CallDataSchema } from './providers/transactions/types.js';
 export {
   randomAddress,
+  randomCosmosAddress,
   randomHookConfig,
   randomIsmConfig,
 } from './test/testUtils.js';
@@ -419,6 +427,7 @@ export {
   EV5GnosisSafeTxSubmitterPropsSchema,
   EV5ImpersonatedAccountTxSubmitterProps,
   EV5ImpersonatedAccountTxSubmitterPropsSchema,
+  EvmIcaTxSubmitterProps,
 } from './providers/transactions/submitter/ethersV5/types.js';
 
 export { TxSubmitterBuilder } from './providers/transactions/submitter/builder/TxSubmitterBuilder.js';
@@ -591,6 +600,7 @@ export {
   EvmNativeTokenAdapter,
   EvmTokenAdapter,
   EvmXERC20VSAdapter,
+  EvmXERC20Adapter,
 } from './token/adapters/EvmTokenAdapter.js';
 export {
   IHypTokenAdapter,
@@ -653,6 +663,7 @@ export {
 export { TokenMetadataMap } from './token/TokenMetadataMap.js';
 export {
   EVM_TOKEN_TYPE_TO_STANDARD,
+  LOCKBOX_STANDARDS,
   MINT_LIMITED_STANDARDS,
   PROTOCOL_TO_NATIVE_STANDARD,
   TOKEN_COLLATERALIZED_STANDARDS,
@@ -705,7 +716,8 @@ export {
   WarpRouteDeployConfigMailboxRequiredSchema,
   WarpRouteDeployConfigSchema,
   WarpRouteDeployConfigSchemaErrors,
-  XERC20LimitConfig,
+  XERC20VSLimitConfig,
+  XERC20Type,
   XERC20TokenExtraBridgesLimits,
   XERC20TokenMetadata,
 } from './token/types.js';
@@ -724,6 +736,8 @@ export {
   PausableConfig,
   PausableSchema,
   ProtocolMap,
+  TypedSigner,
+  IMultiProtocolSignerManager,
 } from './types.js';
 export { getCosmosRegistryChain } from './utils/cosmos.js';
 export { verifyScale } from './utils/decimals.js';
@@ -778,3 +792,17 @@ export {
   WarpTypedTransaction,
 } from './warp/types.js';
 export { WarpCore, WarpCoreOptions } from './warp/WarpCore.js';
+export { EvmTimelockReader } from './timelock/evm/EvmTimelockReader.js';
+export { EvmTimelockDeployer } from './timelock/evm/EvmTimelockDeployer.js';
+export {
+  TimelockConfig,
+  TimelockConfigSchema,
+  TimelockConfigMapSchema,
+  TimelockConfigMap,
+} from './timelock/types.js';
+export {
+  CANCELLER_ROLE,
+  EXECUTOR_ROLE,
+  PROPOSER_ROLE,
+} from './timelock/evm/constants.js';
+export { EvmEventLogsReader } from './rpc/evm/EvmEventLogsReader.js';